+++ docs/Sphinx/source/ref/doxygen/doxygen_html/menudata.js
@@ -138,8 +138,5 @@ var menudata={children:[
 {text:"p",url:"globals_func.html#index_p"},
 {text:"s",url:"globals_func.html#index_s"},
 {text:"t",url:"globals_func.html#index_t"}]},
-{text:"Variables",url:"globals_vars.html",children:[
-{text:"c",url:"globals_vars.html#index_c"},
-{text:"p",url:"globals_vars.html#index_p"},
-{text:"s",url:"globals_vars.html#index_s"}]},
+{text:"Variables",url:"globals_vars.html"},
 {text:"Macros",url:"globals_defs.html"}]}]}]}