+++ docs/Sphinx/source/ref/doxygen/doxygen_html/search/classes_2.js
@@ -1,14 +1,14 @@
 var searchData=
 [
-  ['exceptionsparsevirtualfile_438',['ExceptionSparseVirtualFile',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file.html',1,'SVFS::Exceptions']]],
-  ['exceptionsparsevirtualfilediff_439',['ExceptionSparseVirtualFileDiff',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_diff.html',1,'SVFS::Exceptions']]],
-  ['exceptionsparsevirtualfileerase_440',['ExceptionSparseVirtualFileErase',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_erase.html',1,'SVFS::Exceptions']]],
-  ['exceptionsparsevirtualfileread_441',['ExceptionSparseVirtualFileRead',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_read.html',1,'SVFS::Exceptions']]],
-  ['exceptionsparsevirtualfilesystem_442',['ExceptionSparseVirtualFileSystem',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system.html',1,'SVFS::Exceptions']]],
-  ['exceptionsparsevirtualfilesysteminsert_443',['ExceptionSparseVirtualFileSystemInsert',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_insert.html',1,'SVFS::Exceptions']]],
-  ['exceptionsparsevirtualfilesystemoutofrange_444',['ExceptionSparseVirtualFileSystemOutOfRange',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_out_of_range.html',1,'SVFS::Exceptions']]],
-  ['exceptionsparsevirtualfilesystemremove_445',['ExceptionSparseVirtualFileSystemRemove',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_remove.html',1,'SVFS::Exceptions']]],
-  ['exceptionsparsevirtualfilewrite_446',['ExceptionSparseVirtualFileWrite',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_write.html',1,'SVFS::Exceptions']]],
-  ['exceptiontest_447',['ExceptionTest',['../class_s_v_f_s_1_1_test_1_1_exception_test.html',1,'SVFS::Test']]],
-  ['exceptiontestconfiguration_448',['ExceptionTestConfiguration',['../class_s_v_f_s_1_1_test_1_1_exception_test_configuration.html',1,'SVFS::Test']]]
+  ['exceptionsparsevirtualfile_387',['ExceptionSparseVirtualFile',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file.html',1,'SVFS::Exceptions']]],
+  ['exceptionsparsevirtualfilediff_388',['ExceptionSparseVirtualFileDiff',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_diff.html',1,'SVFS::Exceptions']]],
+  ['exceptionsparsevirtualfileerase_389',['ExceptionSparseVirtualFileErase',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_erase.html',1,'SVFS::Exceptions']]],
+  ['exceptionsparsevirtualfileread_390',['ExceptionSparseVirtualFileRead',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_read.html',1,'SVFS::Exceptions']]],
+  ['exceptionsparsevirtualfilesystem_391',['ExceptionSparseVirtualFileSystem',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system.html',1,'SVFS::Exceptions']]],
+  ['exceptionsparsevirtualfilesysteminsert_392',['ExceptionSparseVirtualFileSystemInsert',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_insert.html',1,'SVFS::Exceptions']]],
+  ['exceptionsparsevirtualfilesystemoutofrange_393',['ExceptionSparseVirtualFileSystemOutOfRange',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_out_of_range.html',1,'SVFS::Exceptions']]],
+  ['exceptionsparsevirtualfilesystemremove_394',['ExceptionSparseVirtualFileSystemRemove',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_remove.html',1,'SVFS::Exceptions']]],
+  ['exceptionsparsevirtualfilewrite_395',['ExceptionSparseVirtualFileWrite',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_write.html',1,'SVFS::Exceptions']]],
+  ['exceptiontest_396',['ExceptionTest',['../class_s_v_f_s_1_1_test_1_1_exception_test.html',1,'SVFS::Test']]],
+  ['exceptiontestconfiguration_397',['ExceptionTestConfiguration',['../class_s_v_f_s_1_1_test_1_1_exception_test_configuration.html',1,'SVFS::Test']]]
 ];