+++ docs/Sphinx/source/ref/doxygen/doxygen_html/search/functions_2.js
@@ -1,9 +1,9 @@
 var searchData=
 [
-  ['block_5fsize_521',['block_size',['../class_s_v_f_s_1_1_sparse_virtual_file.html#aac54ea133274f421d2159a5b4a30641c',1,'SVFS::SparseVirtualFile']]],
-  ['block_5ftouch_522',['block_touch',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a570900eea55726be32f9a0df9d8849e2',1,'SVFS::SparseVirtualFile']]],
-  ['block_5ftouches_523',['block_touches',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a3768e98b92d53dc55298e13d4d7046cb',1,'SVFS::SparseVirtualFile']]],
-  ['blocks_524',['blocks',['../class_s_v_f_s_1_1_sparse_virtual_file.html#aebcb4ccea430a954d95502399306e771',1,'SVFS::SparseVirtualFile']]],
-  ['bytes_5fread_525',['bytes_read',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a009ac567b03d6496a26d3f64de9ff66d',1,'SVFS::SparseVirtualFile']]],
-  ['bytes_5fwrite_526',['bytes_write',['../class_s_v_f_s_1_1_sparse_virtual_file.html#aafa9a8b3b6d90459a5c6cf2eaba772e7',1,'SVFS::SparseVirtualFile']]]
+  ['block_5fsize_470',['block_size',['../class_s_v_f_s_1_1_sparse_virtual_file.html#aac54ea133274f421d2159a5b4a30641c',1,'SVFS::SparseVirtualFile']]],
+  ['block_5ftouch_471',['block_touch',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a570900eea55726be32f9a0df9d8849e2',1,'SVFS::SparseVirtualFile']]],
+  ['block_5ftouches_472',['block_touches',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a3768e98b92d53dc55298e13d4d7046cb',1,'SVFS::SparseVirtualFile']]],
+  ['blocks_473',['blocks',['../class_s_v_f_s_1_1_sparse_virtual_file.html#aebcb4ccea430a954d95502399306e771',1,'SVFS::SparseVirtualFile']]],
+  ['bytes_5fread_474',['bytes_read',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a009ac567b03d6496a26d3f64de9ff66d',1,'SVFS::SparseVirtualFile']]],
+  ['bytes_5fwrite_475',['bytes_write',['../class_s_v_f_s_1_1_sparse_virtual_file.html#aafa9a8b3b6d90459a5c6cf2eaba772e7',1,'SVFS::SparseVirtualFile']]]
 ];