+++ docs/Sphinx/source/ref/doxygen/doxygen_html/search/functions_5.js
@@ -1,17 +1,17 @@
 var searchData=
 [
-  ['erase_578',['erase',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a92ffb36bd3061ea6e08fe2bbb72b5e32',1,'SVFS::SparseVirtualFile']]],
-  ['error_5fmessage_579',['error_message',['../class_s_v_f_s_1_1_test_1_1_test_result.html#aaf914185844834ed96500d4d5062dc26',1,'SVFS::Test::TestResult']]],
-  ['exceptionsparsevirtualfile_580',['ExceptionSparseVirtualFile',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file.html#a279a193af59c819fdd264317bea5b115',1,'SVFS::Exceptions::ExceptionSparseVirtualFile']]],
-  ['exceptionsparsevirtualfilediff_581',['ExceptionSparseVirtualFileDiff',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_diff.html#a03a122511a909900f2426316d2aed121',1,'SVFS::Exceptions::ExceptionSparseVirtualFileDiff']]],
-  ['exceptionsparsevirtualfileerase_582',['ExceptionSparseVirtualFileErase',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_erase.html#a7c53a4ea511cf3adb24188d39f67edb4',1,'SVFS::Exceptions::ExceptionSparseVirtualFileErase']]],
-  ['exceptionsparsevirtualfileread_583',['ExceptionSparseVirtualFileRead',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_read.html#ac67f98f37550883005eddc65a4c7cb74',1,'SVFS::Exceptions::ExceptionSparseVirtualFileRead']]],
-  ['exceptionsparsevirtualfilesystem_584',['ExceptionSparseVirtualFileSystem',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system.html#ac0b6373a246beb7e3c72e593cf4049e2',1,'SVFS::Exceptions::ExceptionSparseVirtualFileSystem']]],
-  ['exceptionsparsevirtualfilesysteminsert_585',['ExceptionSparseVirtualFileSystemInsert',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_insert.html#a461f5e5679540a1c668bff1bca61633c',1,'SVFS::Exceptions::ExceptionSparseVirtualFileSystemInsert']]],
-  ['exceptionsparsevirtualfilesystemoutofrange_586',['ExceptionSparseVirtualFileSystemOutOfRange',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_out_of_range.html#a6bde2274d34328fee58c0d7aefcd4f71',1,'SVFS::Exceptions::ExceptionSparseVirtualFileSystemOutOfRange']]],
-  ['exceptionsparsevirtualfilesystemremove_587',['ExceptionSparseVirtualFileSystemRemove',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_remove.html#aacfa9312e86bfe1c713bbcbf61c951ba',1,'SVFS::Exceptions::ExceptionSparseVirtualFileSystemRemove']]],
-  ['exceptionsparsevirtualfilewrite_588',['ExceptionSparseVirtualFileWrite',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_write.html#a08dab27452ac248ff0b6a214f9d05e8e',1,'SVFS::Exceptions::ExceptionSparseVirtualFileWrite']]],
-  ['exceptiontest_589',['ExceptionTest',['../class_s_v_f_s_1_1_test_1_1_exception_test.html#a9339ac28b9aa9ac15584291c122da070',1,'SVFS::Test::ExceptionTest']]],
-  ['exceptiontestconfiguration_590',['ExceptionTestConfiguration',['../class_s_v_f_s_1_1_test_1_1_exception_test_configuration.html#a70fbfa6f11e2dec24e375e53b730eef0',1,'SVFS::Test::ExceptionTestConfiguration']]],
-  ['exec_5ftime_591',['exec_time',['../class_s_v_f_s_1_1_test_1_1_test_result.html#a01616f9ceff167f0291ea5e2d18189b7',1,'SVFS::Test::TestResult']]]
+  ['erase_525',['erase',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a92ffb36bd3061ea6e08fe2bbb72b5e32',1,'SVFS::SparseVirtualFile']]],
+  ['error_5fmessage_526',['error_message',['../class_s_v_f_s_1_1_test_1_1_test_result.html#aaf914185844834ed96500d4d5062dc26',1,'SVFS::Test::TestResult']]],
+  ['exceptionsparsevirtualfile_527',['ExceptionSparseVirtualFile',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file.html#a279a193af59c819fdd264317bea5b115',1,'SVFS::Exceptions::ExceptionSparseVirtualFile']]],
+  ['exceptionsparsevirtualfilediff_528',['ExceptionSparseVirtualFileDiff',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_diff.html#a03a122511a909900f2426316d2aed121',1,'SVFS::Exceptions::ExceptionSparseVirtualFileDiff']]],
+  ['exceptionsparsevirtualfileerase_529',['ExceptionSparseVirtualFileErase',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_erase.html#a7c53a4ea511cf3adb24188d39f67edb4',1,'SVFS::Exceptions::ExceptionSparseVirtualFileErase']]],
+  ['exceptionsparsevirtualfileread_530',['ExceptionSparseVirtualFileRead',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_read.html#ac67f98f37550883005eddc65a4c7cb74',1,'SVFS::Exceptions::ExceptionSparseVirtualFileRead']]],
+  ['exceptionsparsevirtualfilesystem_531',['ExceptionSparseVirtualFileSystem',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system.html#ac0b6373a246beb7e3c72e593cf4049e2',1,'SVFS::Exceptions::ExceptionSparseVirtualFileSystem']]],
+  ['exceptionsparsevirtualfilesysteminsert_532',['ExceptionSparseVirtualFileSystemInsert',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_insert.html#a461f5e5679540a1c668bff1bca61633c',1,'SVFS::Exceptions::ExceptionSparseVirtualFileSystemInsert']]],
+  ['exceptionsparsevirtualfilesystemoutofrange_533',['ExceptionSparseVirtualFileSystemOutOfRange',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_out_of_range.html#a6bde2274d34328fee58c0d7aefcd4f71',1,'SVFS::Exceptions::ExceptionSparseVirtualFileSystemOutOfRange']]],
+  ['exceptionsparsevirtualfilesystemremove_534',['ExceptionSparseVirtualFileSystemRemove',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_system_remove.html#aacfa9312e86bfe1c713bbcbf61c951ba',1,'SVFS::Exceptions::ExceptionSparseVirtualFileSystemRemove']]],
+  ['exceptionsparsevirtualfilewrite_535',['ExceptionSparseVirtualFileWrite',['../class_s_v_f_s_1_1_exceptions_1_1_exception_sparse_virtual_file_write.html#a08dab27452ac248ff0b6a214f9d05e8e',1,'SVFS::Exceptions::ExceptionSparseVirtualFileWrite']]],
+  ['exceptiontest_536',['ExceptionTest',['../class_s_v_f_s_1_1_test_1_1_exception_test.html#a9339ac28b9aa9ac15584291c122da070',1,'SVFS::Test::ExceptionTest']]],
+  ['exceptiontestconfiguration_537',['ExceptionTestConfiguration',['../class_s_v_f_s_1_1_test_1_1_exception_test_configuration.html#a70fbfa6f11e2dec24e375e53b730eef0',1,'SVFS::Test::ExceptionTestConfiguration']]],
+  ['exec_5ftime_538',['exec_time',['../class_s_v_f_s_1_1_test_1_1_test_result.html#a01616f9ceff167f0291ea5e2d18189b7',1,'SVFS::Test::TestResult']]]
 ];