+++ docs/Sphinx/source/ref/doxygen/doxygen_html/search/functions_d.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['need_610',['need',['../class_s_v_f_s_1_1_sparse_virtual_file.html#ade33f55fdcee892fbc10a44aa219b6c5',1,'SVFS::SparseVirtualFile']]],
-  ['num_5fblocks_611',['num_blocks',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a4031e0bf4b691bc702086bf2403a0777',1,'SVFS::SparseVirtualFile::num_blocks()'],['../class_s_v_f_s_1_1_sparse_virtual_file_system.html#a532f9e1a9978043b641984151ddd93a2',1,'SVFS::SparseVirtualFileSystem::num_blocks()']]],
-  ['num_5fbytes_612',['num_bytes',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a480da981319d2a3a7d2ad46ce6055380',1,'SVFS::SparseVirtualFile::num_bytes()'],['../class_s_v_f_s_1_1_sparse_virtual_file_system.html#a5d3bfb953b75d79bac4ecffdbfd69c5c',1,'SVFS::SparseVirtualFileSystem::num_bytes()']]]
+  ['need_557',['need',['../class_s_v_f_s_1_1_sparse_virtual_file.html#ade33f55fdcee892fbc10a44aa219b6c5',1,'SVFS::SparseVirtualFile']]],
+  ['num_5fblocks_558',['num_blocks',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a4031e0bf4b691bc702086bf2403a0777',1,'SVFS::SparseVirtualFile::num_blocks()'],['../class_s_v_f_s_1_1_sparse_virtual_file_system.html#a532f9e1a9978043b641984151ddd93a2',1,'SVFS::SparseVirtualFileSystem::num_blocks()']]],
+  ['num_5fbytes_559',['num_bytes',['../class_s_v_f_s_1_1_sparse_virtual_file.html#a480da981319d2a3a7d2ad46ce6055380',1,'SVFS::SparseVirtualFile::num_bytes()'],['../class_s_v_f_s_1_1_sparse_virtual_file_system.html#a5d3bfb953b75d79bac4ecffdbfd69c5c',1,'SVFS::SparseVirtualFileSystem::num_bytes()']]]
 ];