+++ docs/Sphinx/source/ref/doxygen/doxygen_html/search/variables_b.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['overwrite_5fchar_835',['OVERWRITE_CHAR',['../namespace_s_v_f_s.html#a1b76300f53102fd75930d04ced854b5d',1,'SVFS']]],
-  ['overwrite_5fon_5fexit_836',['overwrite_on_exit',['../struct_s_v_f_s_1_1_sparse_virtual_file_config.html#ab66d1eed66f3bb7d967790f75ab10bd8',1,'SVFS::SparseVirtualFileConfig']]]
+  ['overwrite_5fchar_735',['OVERWRITE_CHAR',['../namespace_s_v_f_s.html#a1b76300f53102fd75930d04ced854b5d',1,'SVFS']]],
+  ['overwrite_5fon_5fexit_736',['overwrite_on_exit',['../struct_s_v_f_s_1_1_sparse_virtual_file_config.html#ab66d1eed66f3bb7d967790f75ab10bd8',1,'SVFS::SparseVirtualFileConfig']]]
 ];