+++ tests/unit/test_svf_memory.py
@@ -28,7 +28,6 @@ import pytest
 
 import svfsc
 
-
 # 1000 * 10000 * 1000 is around 10Gb
 SLOW_REPEAT = 1000
 SLOW_BLOCKS = 10000
@@ -72,7 +71,7 @@ def test_memory_SVF_need():
         # Not coalesced
         s.write(block_index * 2 * SLOW_BLOCK_SIZE, b' ' * SLOW_BLOCK_SIZE)
     for repeat in range(SLOW_REPEAT * 10):
-        s.need(0, 1024**3)
+        s.need(0, 1024 ** 3)
     rss_diff = proc.memory_info().rss - rss_start_overall
     print(
         f'RSS start: {rss_start_overall:,d}'
@@ -169,32 +168,44 @@ def test_memory_SVF_pickle_loads():
     # assert 0
 
 
-# @pytest.mark.slow
 def test_memory_SVF_write_and_punt():
+    """Write a load of blocks (100Mb) then punt down to two blocks (2,000) bytes checking RSS as we go."""
     proc = psutil.Process()
-    # print()
+    rss_start = proc.memory_info().rss
+    print()
+    print(f'RSS start: {rss_start:,d}')
     ID = 'id'
-    rss_start_overall = proc.memory_info().rss
-    for repeat in range(SLOW_REPEAT):
-        # rss_start = proc.memory_info().rss
-        # print(f'RSS start: {proc.memory_info().rss:,d}', end='')
-        s = svfsc.cSVF(ID, 1.0)
-        for block_index in range(SLOW_BLOCKS):
-            # Not coalesced
-            s.write(block_index * 2 * SLOW_BLOCK_SIZE, b' ' * SLOW_BLOCK_SIZE)
-        # print(
-        #     f' SVF: size_of(): {s.size_of():,d}'
-        #     f' blocks {s.num_blocks()}'
-        #     f' ∆: {proc.memory_info().rss - rss_start:,d}'
-        # )
-        # Now punt
-        s.lru_punt(1024)
-    rss_diff = proc.memory_info().rss - rss_start_overall
+    svf = svfsc.cSVF(ID, 1.0)
+    size_of = svf.size_of()
+    for block_index in range(SLOW_BLOCKS * 10):
+        # Not coalesced
+        svf.write(block_index * 2 * SLOW_BLOCK_SIZE, b' ' * SLOW_BLOCK_SIZE)
+    d_rss = proc.memory_info().rss - rss_start
     print(
-        f'RSS start: {rss_start_overall:,d}'
-        f' end: {proc.memory_info().rss:,d}'
-        f' ∆: {rss_diff:+,d}')
-    assert rss_diff < 40e6
-    assert 0
-
-
+        f'Before punt. SVF: size_of(): {svf.size_of():12,d}'
+        f' ∆: {svf.size_of() - size_of:+12,d}'
+        f' blocks {svf.num_blocks():8,d}'
+        f' bytes {svf.num_bytes():12,d}'
+        f' ∆RSS: {d_rss:+12,d}'
+    )
+    # Check that delta RSS is less than the stored bytes +20MB
+    assert d_rss < (svf.num_bytes() + 20_000_000)
+    # Now punt down to two blocks, 2000 bytes.
+    rss = proc.memory_info().rss
+    size_of = svf.size_of()
+    svf.lru_punt(2 * SLOW_BLOCK_SIZE + 1)
+    d_rss = proc.memory_info().rss - rss
+    print(
+        f'After punt.  SVF: size_of(): {svf.size_of():12,d}'
+        f' ∆: {svf.size_of() - size_of:+12,d}'
+        f' blocks {svf.num_blocks():8,d}'
+        f' bytes {svf.num_bytes():12,d}'
+        f' ∆RSS: {d_rss:+12,d}'
+    )
+    d_rss = proc.memory_info().rss - rss_start
+    assert d_rss < (svf.num_bytes() + 40_000_000)
+    print(
+        f'RSS start: {rss_start:12,d}'
+        f' end: {proc.memory_info().rss:12,d}'
+        f' ∆RSS: {d_rss:+12,d}')
+    assert d_rss < 40e6