+++ tests/unit/test_svfs.py
@@ -109,29 +109,29 @@ def test_SVFS_has(insert, expected_keys):
 @pytest.mark.parametrize(
     'blocks, expected_blocks',
     (
-        # No inserts
-        (
-            tuple(), tuple(),
-        ),
-        # Single insert
-        (
+            # No inserts
             (
-                (0, b' '),
+                    tuple(), tuple(),
             ),
+            # Single insert
             (
-                (0, 1),
+                    (
+                            (0, b' '),
+                    ),
+                    (
+                            (0, 1),
+                    ),
             ),
-        ),
-        # Duplicate insert
-        (
+            # Duplicate insert
             (
-                (0, b' '),
-                (0, b' '),
+                    (
+                            (0, b' '),
+                            (0, b' '),
+                    ),
+                    (
+                            (0, 1),
+                    ),
             ),
-            (
-                (0, 1),
-            ),
-        ),
     )
 )
 def test_SVFS_write(blocks, expected_blocks):
@@ -166,7 +166,7 @@ def test_SVFS_write(blocks, expected_blocks):
     (1, 2, 4, 8, 16, 32, 64),
 )
 def test_simulate_write_uncoalesced(block_size):
-    SIZE = 1024 #* 1024 * 1
+    SIZE = 1024  # * 1024 * 1
     ID = 'abc'
     s = svfsc.cSVFS()
     s.insert(ID, 12.0)
@@ -186,7 +186,7 @@ def test_simulate_write_uncoalesced(block_size):
     (1, 2, 4, 8, 16, 32, 64),
 )
 def test_simulate_write_coalesced(block_size):
-    SIZE = 1024 #* 1024 * 1
+    SIZE = 1024  # * 1024 * 1
     ID = 'abc'
     s = svfsc.cSVFS()
     s.insert(ID, 12.0)
@@ -337,7 +337,7 @@ def test_SVFS_ctor_config(args, kwargs, expected):
     # assert 0
 
 
-def test_SVFS_lru_punt_strategy_builtin():
+def test_SVFS_lru_punt():
     """Example of a LRU cache punting strategy as given."""
     # svf = svfsc.cSVF('id', 1.0)
     svfs = svfsc.cSVFS()
@@ -356,6 +356,31 @@ def test_SVFS_lru_punt_strategy_builtin():
     cache_upper_bound = 1024
     assert svfs.num_bytes(ID) >= cache_upper_bound
     removed = svfs.lru_punt(ID, cache_upper_bound)
+    assert removed == (block_size * block_count - 896)
+    assert svfs.num_bytes(ID) < cache_upper_bound
+    assert svfs.num_bytes(ID) == 896
+    assert svfs.num_blocks(ID) == 896 // block_size
+
+
+def test_SVFS_lru_punt_all():
+    """Near duplicate of test_SVFS_lru_punt()."""
+    # svf = svfsc.cSVF('id', 1.0)
+    svfs = svfsc.cSVFS()
+    ID = 'abc'
+    svfs.insert(ID, 1.0)
+    fpos = 0
+    block_size = 128
+    block_count = 256
+    for i in range(block_count):
+        svfs.write(ID, fpos, b' ' * block_size)
+        fpos += block_size
+        fpos += 1
+    assert svfs.num_bytes(ID) == block_count * block_size
+    assert svfs.num_blocks(ID) == block_count
+    assert len(svfs.block_touches(ID)) == block_count
+    cache_upper_bound = 1024
+    assert svfs.num_bytes(ID) >= cache_upper_bound
+    removed = svfs.lru_punt_all(cache_upper_bound)
     assert removed == (block_size * block_count - 896)
     assert svfs.num_bytes(ID) < cache_upper_bound
     assert svfs.num_bytes(ID) == 896