+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 =====================
 
 - Add cache punting.
+- Make C docstrings type parsable (good for Sphinx) and add a script that can create a mypy stub file.
 - Development Status :: 4 - Beta
 
 0.1.2 (2023-10-03)